backend/requirements.txt
fastapi
uvicorn[standard]
asyncpg
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
pydantic
pydantic-settings
email-validator==2.1.0
pywebpush
cryptography
xhtml2pdf
jinja2
